+++ bundles/AdminBundle/Resources/public/css/icons.css
@@ -1436,8 +1436,8 @@
     background: url(/bundles/pimcoreadmin/img/flat-color-icons/neutral_trading.svg) center center no-repeat !important;
 }
 
-.pimcore_icon_reverseManyToManyObjectRelation {
-    background: url(/bundles/pimcoreadmin/img/flat-color-icons/reverse_many_to_many_object_relation.svg) center center no-repeat !important;
+.pimcore_icon_reverseObjectRelation {
+    background: url(/bundles/pimcoreadmin/img/flat-color-icons/reverse_object_relation.svg) center center no-repeat !important;
 }
 
 .pimcore_icon_manyToManyObjectRelation {
