+++ bundles/AdminBundle/Resources/public/js/pimcore/object/objectbricks/field.js
@@ -16,7 +16,7 @@ pimcore.object.objectbricks.field = Class.create(pimcore.object.classes.klass, {
 
     allowedInType: 'objectbrick',
     disallowedDataTypes: [
-        "reverseManyToManyObjectRelation",
+        "reverseObjectRelation",
         "user",
         "fieldcollections",
         "localizedfields",
