+++ bundles/AdminBundle/Resources/public/js/pimcore/object/tags/reverseObjectRelation.js
@@ -11,7 +11,7 @@
  * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
 
-pimcore.registerNS("pimcore.object.tags.reverseManyToManyObjectRelation");
+pimcore.registerNS("pimcore.object.tags.reverseObjectRelation");
 pimcore.object.tags.reverseManyToManyObjectRelation = Class.create(pimcore.object.tags.manyToManyObjectRelation, {
 
     pathProperty: "path",
@@ -425,3 +425,4 @@ pimcore.object.tags.reverseManyToManyObjectRelation = Class.create(pimcore.objec
 
 // @TODO BC layer, to be removed in Pimcore 10
 pimcore.object.tags.nonownerobjects = pimcore.object.tags.reverseManyToManyObjectRelation;
+pimcore.object.tags.reverseManyToManyObjectRelation = pimcore.object.tags.reverseObjectRelation;
