+++ bundles/AdminBundle/Resources/views/Admin/Index/index.html.php
@@ -490,7 +490,7 @@ $scripts = array(
     "pimcore/object/classes/data/manyToManyObjectRelation.js",
     "pimcore/object/classes/data/advancedManyToManyRelation.js",
     "pimcore/object/classes/data/advancedManyToManyObjectRelation.js",
-    "pimcore/object/classes/data/reverseManyToManyObjectRelation.js",
+    "pimcore/object/classes/data/reverseObjectRelation.js",
     "pimcore/object/classes/data/booleanSelect.js",
     "pimcore/object/classes/data/select.js",
     "pimcore/object/classes/data/urlSlug.js",
@@ -614,7 +614,7 @@ $scripts = array(
     "pimcore/object/gridcolumn/operator/FieldCollectionGetter.js",
     "pimcore/object/gridcolumn/operator/ObjectBrickGetter.js",
     "pimcore/object/tags/advancedManyToManyObjectRelation.js",
-    "pimcore/object/tags/reverseManyToManyObjectRelation.js",
+    "pimcore/object/tags/reverseObjectRelation.js",
     "pimcore/object/tags/urlSlug.js",
     "pimcore/object/tags/booleanSelect.js",
     "pimcore/object/tags/select.js",
