+++ bundles/CoreBundle/Resources/translations/aliases.json
@@ -134,7 +134,8 @@
   "multihref": "many_to_many_relation",
   "multihrefMetadata": "advanced_many_to_many_relation",
   "objectsMetadata": "advanced_many_to_many_object_relation",
-  "nonownerobjects": "reverse_many_to_many_object_relation",
+  "nonownerobjects": "reverse_object_relation",
+  "reverseManyToManyObjectRelation": "reverse_object_relation",
   "href": "many_to_one_relation",
   "bool": "boolean"
 }
