+++ models/DataObject/Concrete/Dao.php
@@ -201,7 +201,7 @@ class Dao extends Model\DataObject\AbstractObject\Dao
         foreach ($fieldDefinitions as $key => $fd) {
             if (($fd instanceof LazyLoadingSupportInterface || method_exists($fd, 'getLazyLoading'))
                                     && $fd->getLazyLoading()) {
-                if (!$this->model->isLazyKeyLoaded($key) || $fd instanceof DataObject\ClassDefinition\Data\ReverseManyToManyObjectRelation) {
+                if (!$this->model->isLazyKeyLoaded($key) || $fd instanceof DataObject\ClassDefinition\Data\ReverseObjectRelation) {
                     //this is a relation subject to lazy loading - it has not been loaded
                     $untouchable[] = $key;
                 }
