+++ tests/_support/Resources/objects/class-import.json
@@ -3141,7 +3141,7 @@
                                 "visibleSearch": true
                             },
                             {
-                                "fieldtype": "reverseManyToManyObjectRelation",
+                                "fieldtype": "reverseObjectRelation",
                                 "ownerClassName": null,
                                 "ownerClassId": null,
                                 "ownerFieldName": null,
